Establish a dialog upon a 200-class response to a SUBSCRIBE request
In our case, we are not ever going to send a SUBSCRIBE on a
pre-existing dialog, so this covers the standard case for establishing
a new dialog.

What is currently _not_ handled by this is the establishment of a new
dialog if the matching NOTIFY request arrives before the response to
SUBSCRIBE.

See section 3.3.4 of RFC 3265.