Provide the ability to reSUBSCRIBE within an existing dialog
The same event and expires options should be provided as this
information is currently not obtained from the dialog.

This avoids us needing to do any deep surgery on dialogs for the
moment until the new engine is in place.  What a SipApp will need to
do is remember the package and provide it with the re-subscribe.