Callback to allow a SipApp to provide a list of supported features
This callback is primarily used when checking the contents of the
Require header in a request.  If the list of extensions included in
the Require header is a subset of those supported by the SipApp, then
the request will proceed as usual rather than responding with a bad
extension response.